backend/school/requirements/development.txt
# Development requirements

-r base.txt

django-debug-toolbar==3.2.4
pytest-django==4.5.2
mixer==7.2.0
pytest-cov==3.0.0
flake8==4.0.1
flake8-isort==4.1.1
black==21.12b0
pylint-django==2.5.0
pre-commit==2.16.0
factory-boy==3.2.1
django-extensions==3.1.5
Pillow==9.1.0
djangorestframework-simplejwt==5.1.0
pyjwt==2.3.0
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
phonenumbers==8.12.48
